+++ realm/realm-library/src/objectServer/java/io/realm/SyncSession.java
@@ -45,7 +45,7 @@ import io.realm.internal.Util;
 import io.realm.internal.android.AndroidCapabilities;
 import io.realm.internal.async.RealmAsyncTaskImpl;
 import io.realm.internal.network.AuthenticateResponse;
-import io.realm.internal.network.AuthenticationServer;
+import io.realm.internal.network.RealmObjectServer;
 import io.realm.internal.network.ExponentialBackoffTask;
 import io.realm.internal.network.NetworkStateReceiver;
 import io.realm.internal.objectserver.Token;
@@ -741,7 +741,7 @@ public class SyncSession {
     }
 
     // Return the access token for the Realm this Session is connected to.
-    String getAccessToken(final AuthenticationServer authServer, String refreshToken) {
+    String getAccessToken(final RealmObjectServer authServer, String refreshToken) {
         // check first if there's a valid access_token we can return immediately
         if (getUser().isRealmAuthenticated(configuration)) {
             Token accessToken = getUser().getAccessToken(configuration);
@@ -773,7 +773,7 @@ public class SyncSession {
     }
 
     // Authenticate by getting access tokens for the specific Realm
-    private void authenticateRealm(final AuthenticationServer authServer) {
+    private void authenticateRealm(final RealmObjectServer authServer) {
         if (networkRequest != null) {
             networkRequest.cancel();
         }
@@ -828,7 +828,7 @@ public class SyncSession {
         networkRequest = new RealmAsyncTaskImpl(task, SyncManager.NETWORK_POOL_EXECUTOR);
     }
 
-    private void scheduleRefreshAccessToken(final AuthenticationServer authServer, long expireDateInMs) {
+    private void scheduleRefreshAccessToken(final RealmObjectServer authServer, long expireDateInMs) {
         onGoingAccessTokenQuery.set(true);
         // calculate the delay time before which we should refresh the access_token,
         // we adjust to 10 second to proactively refresh the access_token before the session
@@ -862,7 +862,7 @@ public class SyncSession {
     }
 
     // Authenticate by getting access tokens for the specific Realm
-    private void refreshAccessToken(final AuthenticationServer authServer) {
+    private void refreshAccessToken(final RealmObjectServer authServer) {
         // Authenticate in a background thread. This allows incremental backoff and retries in a safe manner.
         clearScheduledAccessTokenRefresh();